avoid using io.Copy
io.Copy treats EOF as a common scenario and never returns it. we are
interested in detecting EOF since it means the piped program exited
and closed it's standard output fd.

We think there were some scenarios where io.Copy was returning 0 bytes
read that were causing the program to exit and then crashing the piped
program.